requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.0
    # via streamlit
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via streamlit
faiss-cpu==1.12.0
    # via basic-rag-app (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
narwhals==2.5.0
    # via altair
numpy==2.3.3
    # via
    #   basic-rag-app (pyproject.toml)
    #   faiss-cpu
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   faiss-cpu
    #   streamlit
pandas==2.3.2
    # via streamlit
pillow==11.3.0
    # via streamlit
protobuf==6.32.1
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via basic-rag-app (pyproject.toml)
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   basic-rag-app (pyproject.toml)
    #   streamlit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.49.1
    # via basic-rag-app (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   streamlit
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
